requirements.txt
Flask==3.0.0
gunicorn==20.1.0
google-auth==2.32.0
google-cloud-aiplatform==1.59.0